Added conditional statement to make or preserve image directory

+++ WMS_API_Tool/run.py
@@ -33,7 +33,7 @@ def main():
                ]
     dates = ["2019-09-12","2020-09-12","2021-09-12","2022-09-12"]
     #Note, Imgdir_make includes the WMS Req function.    
-    #layer_pull(satlist, dates[1::], 'World')
+    layer_pull(satlist, dates[1::], 'World')
     #print(satlist[0].layer_attr(satlist[0].wms))
     
 def layer_pull(satname, date, region):
@@ -42,16 +42,29 @@ def layer_pull(satname, date, region):
         dates = date
         pri_dir = "Image_Directory"
         os.chdir("WMS_API_Tool")
-        os.chdir(pri_dir)
-        for d in range(0, len(dates)):
-            pathname = region + '_' + dates[d]
-            os.makedirs(pathname)
-            os.chdir(pathname)
-            for s in range(0, len(satname)):
-                img = satname[s].wms_req(dates[d])
-                with open(satname[s].abr + "_" + dates[d] + '.png', 'wb') as out:
-                    out.write(img.read())
-            os.chdir('..')      
+        if os.path.exists(pri_dir):        
+            os.chdir(pri_dir)
+            for d in range(0, len(dates)):
+                pathname = region + '_' + dates[d]
+                os.makedirs(pathname)
+                os.chdir(pathname)
+                for s in range(0, len(satname)):
+                    img = satname[s].wms_req(dates[d])
+                    with open(satname[s].abr + "_" + dates[d] + '.png', 'wb') as out:
+                        out.write(img.read())
+                os.chdir('..')
+        else:
+            os.mkdir(pri_dir)
+            os.chdir(pri_dir)
+            for d in range(0, len(dates)):
+                pathname = region + '_' + dates[d]
+                os.makedirs(pathname)
+                os.chdir(pathname)
+                for s in range(0, len(satname)):
+                    img = satname[s].wms_req(dates[d])
+                    with open(satname[s].abr + "_" + dates[d] + '.png', 'wb') as out:
+                        out.write(img.read())
+                os.chdir('..')
     elif isinstance(date, list):
         dates = date
         os.chdir("WMS_API_Tool")
